+++ src/main/java/com/ontotext/refine/client/command/rdf/OutputType.java
@@ -1,0 +1,10 @@
+package com.ontotext.refine.client.command.rdf;
+
+/**
+ * Provides the available types for the result output.
+ *
+ * @author Antoniy Kunchev
+ */
+enum OutputType {
+  STRING, FILE
+}