+++ src/main/java/com/ontotext/refine/client/command/rdf/ResultFormat.java
@@ -39,38 +39,12 @@ public enum ResultFormat {
   HDT(RDFFormat.HDT);
 
   private final RDFFormat rdfFormat;
-  private ResultType as;
 
   private ResultFormat(RDFFormat format) {
     this.rdfFormat = format;
-    this.as = ResultType.STRING;
   }
 
   public RDFFormat getRdfFormat() {
     return rdfFormat;
-  }
-
-  /**
-   * Sets the expected result type.
-   *
-   * @param as the type of the result
-   * @return current {@link ResultFormat} object
-   */
-  public ResultFormat as(ResultType as) {
-    this.as = as;
-    return this;
-  }
-
-  public ResultType getAs() {
-    return as;
-  }
-
-  /**
-   * Provides options for type of the result in which the exported data can be returned.
-   *
-   * @author Antoniy Kunchev
-   */
-  public enum ResultType {
-    STRING, FILE, STREAM
   }
 }